# docs/swagger.yaml
basePath: /
definitions:
  dto.Actor:
    properties:
      actorName:
        type: string
    required:
    - actorName
    type: object
...
  dto.Director:
    properties:
      directorName:
        type: string
    required:
    - directorName
    type: object
...
  utils.Response:
    properties:
      error: {}
      message:
        type: string
      results: {}
      success:
        type: boolean
    type: object
host: localhost:8888
info:
  contact: {}
  description: This is a simple movie ticketing API
  title: Be-Tickitz API
  version: "1.0"
paths:
  /actors:
    get:
      description: Retrieve all actors
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get all actors
      tags:
      - Actors
  /admin/actors:
    post:
      consumes:
      - application/json
      description: Admin only. Add a new actor
      parameters:
      - description: Actor data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.Actor'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Create actor
      tags:
      - Actors
  /admin/actors/{id}:
    delete:
      description: Admin only. Delete a actor by ID
      parameters:
      - description: Actor ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Delete a actor
      tags:
      - Actors
  /admin/directors:
    post:
      consumes:
      - application/json
      description: Admin only. Add a new director
      parameters:
      - description: Director data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.Director'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Create director
      tags:
      - Directors
  /admin/directors/{id}:
    delete:
      description: Admin only. Delete a director by ID
      parameters:
      - description: Director ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Delete a director
      tags:
      - Directors
...
  /directors:
    get:
      description: Retrieve all directors
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get all directors
      tags:
      - Directors
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"